chore: pull up pattern and add class State
